Integrated local song access via songs.xml into the app

+++ shuffle-my-music-app/app/src/main/java/com/frozen_foo/shuffle_my_music_app/io/local/LocalDirectoryAccess.java
@@ -20,7 +20,9 @@ public class LocalDirectoryAccess {
 
 	public void cleanLocalDir() throws IOException {
 		File shuffleMyMusicDir = localDir();
-		FileUtils.cleanDirectory(shuffleMyMusicDir);
+		if (shuffleMyMusicDir.exists()) {
+			FileUtils.cleanDirectory(shuffleMyMusicDir);
+		}
 	}
 
 	public void copyToLocal(InputStream source, String fileName) throws IOException {
